Removes unnecesary false constant
